requirements.txt
amqp==2.5.2
asgiref==3.2.3
autopep8==1.5
beautifulsoup4==4.8.2
billiard==3.6.3.0
celery==4.4.1
certifi==2019.11.28
cfgv==3.1.0
chardet==3.0.4
coverage==5.0.3
coveralls==1.10.0
cssutils==1.0.2
dj-database-url==0.5.0
Django==3.0.2
django-cors-middleware==1.5.0
django-extensions==2.2.6
django-filter==2.2.0
django-heroku==0.3.1
django-inlinecss==0.3.0
django-redis==4.11.0
django-silk==4.0.1
djangorestframework==3.11.0
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
future==0.18.2
gprof2dot==2019.11.30
identify==1.4.11
idna==2.8
importlib-metadata==1.5.0
Jinja2==2.11.1
kombu==4.6.8
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
nodeenv==1.3.5
packaging==20.1
pluggy==0.13.1
pre-commit==2.1.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pynliner==0.8.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.3
PyYAML==5.3
redis==3.4.1
requests==2.22.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
toml==0.10.0
tox==3.14.3
urllib3==1.25.8
vine==1.3.0
virtualenv==16.7.9
whitenoise==5.0.1
zipp==2.1.0
